requirements.txt
google-cloud-aiplatform==1.36.3
